backend/requirements.txt
fastapi
uvicorn
google-genai
pydantic
httpx
elevenlabs
